+++ plugin-dev/Source/Sentry/Private/Interface/SentryLogInterface.h
@@ -1,0 +1,19 @@
+// Copyright (c) 2025 Sentry. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+
+#include "SentryDataTypes.h"
+
+class ISentryLog
+{
+
+public:
+	virtual ~ISentryLog() = default;
+
+	virtual void SetMessage(const FString& message) = 0;
+	virtual FString GetMessage() const = 0;
+	virtual void SetLevel(ESentryLevel level) = 0;
+	virtual ESentryLevel GetLevel() const = 0;
+};