+++ plugin-dev/Source/Sentry/Private/SentryBeforeLogHandler.cpp
@@ -1,0 +1,9 @@
+// Copyright (c) 2025 Sentry. All Rights Reserved.
+
+#include "SentryBeforeLogHandler.h"
+#include "SentryLog.h"
+
+USentryLog* USentryBeforeLogHandler::HandleBeforeLog_Implementation(USentryLog* LogData)
+{
+	return LogData;
+}